Fix for FEC-3276 and FEC-3168.
A very short stream (2 segments) could not be played. The buffer was
longer than the video; this is now capped in VOD cases. In addition,
VOD logic related to stopping stream on very short streams was not
firing properly. This is now fixed. Finally, manifests were being
reloaded on VOD streams, and this is now disabled resulting in faster
startup times and bitrate switches for VOD.